requirements.txt
pandas
scikit-learn 
pillow
torchvision
torch
tqdm